@@ -8,6 +8,7 @@ and value confirmations.
 
 from __future__ import annotations
 
+from abc import ABC
 from typing import Any, Generic, Self, Union
 
 from pydantic import (
@@ -19,7 +20,7 @@ from pydantic import (
 )
 
 from portia.common import SERIALIZABLE_TYPE_VAR, PortiaEnum
-from portia.prefixed_uuid import ClarificationUUID, WorkflowUUID
+from portia.prefixed_uuid import ClarificationUUID, PlanRunUUID
 
 
 class ClarificationCategory(PortiaEnum):
@@ -30,7 +31,6 @@ class ClarificationCategory(PortiaEnum):
     handling and processing.
     """
 
-    ARGUMENT = "Argument"
     ACTION = "Action"
     INPUT = "Input"
     MULTIPLE_CHOICE = "Multiple Choice"
@@ -38,7 +38,7 @@ class ClarificationCategory(PortiaEnum):
     CUSTOM = "Custom"
 
 
-class Clarification(BaseModel, Generic[SERIALIZABLE_TYPE_VAR]):
+class Clarification(BaseModel, Generic[SERIALIZABLE_TYPE_VAR], ABC):
     """Base Model for Clarifications.
 
     A Clarification represents a question or action that requires user input to resolve. For example
@@ -59,8 +59,8 @@ class Clarification(BaseModel, Generic[SERIALIZABLE_TYPE_VAR]):
         default_factory=ClarificationUUID,
         description="A unique ID for this clarification",
     )
-    workflow_id: WorkflowUUID = Field(
-        description="The workflow this clarification is for",
+    plan_run_id: PlanRunUUID = Field(
+        description="The run this clarification is for",
     )
     category: ClarificationCategory = Field(
         description="The category of this clarification",
@@ -76,25 +76,6 @@ class Clarification(BaseModel, Generic[SERIALIZABLE_TYPE_VAR]):
     resolved: bool = Field(
         default=False,
         description="Whether this clarification has been resolved.",
-    )
-
-
-class ArgumentClarification(Clarification[SERIALIZABLE_TYPE_VAR]):
-    """Clarification about a specific argument for a tool.
-
-    This clarification is used when a tool's argument is missing or requires further clarification.
-    The name of the argument is provided within the clarification.
-
-    Attributes:
-        argument_name (str): The name of the argument that is being clarified.
-        category (ClarificationCategory): The category for this clarification, 'Argument'.
-
-    """
-
-    argument_name: str
-    category: ClarificationCategory = Field(
-        default=ClarificationCategory.ARGUMENT,
-        description="The category of this clarification",
     )
 
 
@@ -130,7 +111,7 @@ class ActionClarification(Clarification[SERIALIZABLE_TYPE_VAR]):
         return str(action_url)
 
 
-class InputClarification(ArgumentClarification[SERIALIZABLE_TYPE_VAR]):
+class InputClarification(Clarification[SERIALIZABLE_TYPE_VAR]):
     """Input-based clarification.
 
     Represents a clarification where the user needs to provide a value for a specific argument.
@@ -141,13 +122,16 @@ class InputClarification(ArgumentClarification[SERIALIZABLE_TYPE_VAR]):
 
     """
 
+    argument_name: str = Field(
+        description="The name of the argument that a value is needed for.",
+    )
     category: ClarificationCategory = Field(
         default=ClarificationCategory.INPUT,
         description="The category of this clarification",
     )
 
 
-class MultipleChoiceClarification(ArgumentClarification[SERIALIZABLE_TYPE_VAR]):
+class MultipleChoiceClarification(Clarification[SERIALIZABLE_TYPE_VAR]):
     """Multiple choice-based clarification.
 
     Represents a clarification where the user needs to select an option for a specific argument.
@@ -162,6 +146,9 @@ class MultipleChoiceClarification(ArgumentClarification[SERIALIZABLE_TYPE_VAR]):
 
     """
 
+    argument_name: str = Field(
+        description="The name of the argument that a value is needed for.",
+    )
     category: ClarificationCategory = Field(
         default=ClarificationCategory.MULTIPLE_CHOICE,
         description="The category of this clarification",
@@ -187,7 +174,7 @@ class MultipleChoiceClarification(ArgumentClarification[SERIALIZABLE_TYPE_VAR]):
         return self
 
 
-class ValueConfirmationClarification(ArgumentClarification[SERIALIZABLE_TYPE_VAR]):
+class ValueConfirmationClarification(Clarification[SERIALIZABLE_TYPE_VAR]):
     """Value confirmation clarification.
 
     Represents a clarification where the user is presented with a value and must confirm or deny it.
@@ -199,6 +186,9 @@ class ValueConfirmationClarification(ArgumentClarification[SERIALIZABLE_TYPE_VAR
 
     """
 
+    argument_name: str = Field(
+        description="The name of the argument that whose value needs confirmation.",
+    )
     category: ClarificationCategory = Field(
         default=ClarificationCategory.VALUE_CONFIRMATION,
         description="The category of this clarification",
