@@ -13,7 +13,6 @@ import builtins
 import importlib.metadata
 import json
 import os
-import webbrowser
 from enum import Enum
 from functools import wraps
 from pathlib import Path
@@ -22,23 +21,26 @@ from typing import TYPE_CHECKING, Any, Callable
 import click
 from dotenv import load_dotenv
 from pydantic import BaseModel, Field
+from pydantic_core import PydanticUndefined
 
-from portia.clarification import (
-    ActionClarification,
-    CustomClarification,
-    InputClarification,
-    MultipleChoiceClarification,
-    ValueConfirmationClarification,
-)
-from portia.config import Config, StorageClass
+from portia.clarification_handler import ClarificationHandler
+from portia.config import Config
 from portia.execution_context import execution_context
 from portia.logger import logger
-from portia.runner import Runner
+from portia.portia import ExecutionHooks, Portia
 from portia.tool_registry import DefaultToolRegistry
-from portia.workflow import WorkflowState
 
 if TYPE_CHECKING:
     from pydantic.fields import FieldInfo
+
+    from portia.clarification import (
+        ActionClarification,
+        Clarification,
+        CustomClarification,
+        InputClarification,
+        MultipleChoiceClarification,
+        ValueConfirmationClarification,
+    )
 
 DEFAULT_FILE_PATH = ".portia"
 PORTIA_API_KEY = "portia_api_key"
@@ -85,7 +87,7 @@ class CLIConfig(BaseModel):
     )
 
 
-def generate_cli_option_from_pydantic_field(
+def generate_cli_option_from_pydantic_field(  # noqa: C901
     f: Callable[..., Any],
     field: str,
     info: FieldInfo,
@@ -119,7 +121,12 @@ def generate_cli_option_from_pydantic_field(
                     [e.name for e in info.annotation],
                     case_sensitive=False,
                 )
-                field_default = info.default.name if info.default else None
+                if info.default and info.default is not PydanticUndefined:
+                    field_default = info.default.name
+                elif info.default_factory:
+                    field_default = info.default_factory().name  # type: ignore[reportCallIssue]
+                else:
+                    field_default = None
 
     field_help = info.description or f"Set the value for {option_name}"
 
@@ -145,6 +152,87 @@ def common_options(f: Callable[..., Any]) -> Callable[..., Any]:
     return wrapper
 
 
+class CLIExecutionHooks(ExecutionHooks):
+    """Execution hooks for the CLI."""
+
+    def __init__(self) -> None:
+        """Set up execution hooks for the CLI."""
+        super().__init__(clarification_handler=CLIClarificationHandler())
+
+
+class CLIClarificationHandler(ClarificationHandler):
+    """Handles clarifications by obtaining user input from the CLI."""
+
+    def handle_action_clarification(
+        self,
+        clarification: ActionClarification,
+        on_resolution: Callable[[Clarification, object], None],  # noqa: ARG002
+        on_error: Callable[[Clarification, object], None],  # noqa: ARG002
+    ) -> None:
+        """Handle an action clarification.
+
+        Does this by showing the user the URL on the CLI and instructing them to click on
+        it to proceed.
+        """
+        logger().info(
+            click.style(
+                f"{clarification.user_guidance} -- Please click on the link below to proceed."
+                f"{clarification.action_url}",
+                fg=87,
+            ),
+        )
+
+    def handle_input_clarification(
+        self,
+        clarification: InputClarification,
+        on_resolution: Callable[[Clarification, object], None],
+        on_error: Callable[[Clarification, object], None],  # noqa: ARG002
+    ) -> None:
+        """Handle a user input clarifications by asking the user for input from the CLI."""
+        user_input = click.prompt(
+            click.style(clarification.user_guidance + "\nPlease enter a value:\n", fg=87),
+        )
+        return on_resolution(clarification, user_input)
+
+    def handle_multiple_choice_clarification(
+        self,
+        clarification: MultipleChoiceClarification,
+        on_resolution: Callable[[Clarification, object], None],
+        on_error: Callable[[Clarification, object], None],  # noqa: ARG002
+    ) -> None:
+        """Handle a multi-choice clarification by asking the user for input from the CLI."""
+        choices = click.Choice(clarification.options)
+        user_input = click.prompt(
+            click.style(clarification.user_guidance + "\nPlease choose a value:\n", fg=87),
+            type=choices,
+        )
+        return on_resolution(clarification, user_input)
+
+    def handle_value_confirmation_clarification(
+        self,
+        clarification: ValueConfirmationClarification,
+        on_resolution: Callable[[Clarification, object], None],
+        on_error: Callable[[Clarification, object], None],
+    ) -> None:
+        """Handle a value confirmation clarification by asking the user to confirm from the CLI."""
+        if click.confirm(text=click.style(clarification.user_guidance, fg=87), default=False):
+            on_resolution(clarification, True)  # noqa: FBT003
+        else:
+            on_error(clarification, "Clarification was rejected by the user")
+
+    def handle_custom_clarification(
+        self,
+        clarification: CustomClarification,
+        on_resolution: Callable[[Clarification, object], None],
+        on_error: Callable[[Clarification, object], None],  # noqa: ARG002
+    ) -> None:
+        """Handle a custom clarification."""
+        click.echo(click.style(clarification.user_guidance, fg=87))
+        click.echo(click.style(f"Additional data: {json.dumps(clarification.data)}", fg=87))
+        user_input = click.prompt(click.style("\nPlease enter a value:\n", fg=87))
+        return on_resolution(clarification, user_input)
+
+
 @click.group(context_settings={"max_content_width": 240})
 def cli() -> None:
     """Portia CLI."""
@@ -159,7 +247,7 @@ def version() -> None:
 @click.command()
 @common_options
 @click.argument("query")
-def run(  # noqa: C901
+def run(
     query: str,
     **kwargs,  # noqa: ANN003
 ) -> None:
@@ -170,63 +258,24 @@ def run(  # noqa: C901
     registry = DefaultToolRegistry(config)
 
     # Run the query
-    runner = Runner(
+    portia = Portia(
         config=config,
         tools=(
             registry.match_tools(tool_ids=[cli_config.tool_id]) if cli_config.tool_id else registry
         ),
+        execution_hooks=CLIExecutionHooks(),
     )
 
     with execution_context(end_user_id=cli_config.end_user_id):
-        plan = runner.generate_plan(query)
+        plan = portia.plan(query)
 
         if cli_config.confirm:
             click.echo(plan.model_dump_json(indent=4))
             if not click.confirm("Do you want to execute the plan?"):
                 return
 
-        workflow = runner.create_workflow(plan)
-        workflow = runner.execute_workflow(workflow)
-
-        final_states = [WorkflowState.COMPLETE, WorkflowState.FAILED]
-        while workflow.state not in final_states:
-            for clarification in workflow.get_outstanding_clarifications():
-                if isinstance(clarification, MultipleChoiceClarification):
-                    choices = click.Choice(clarification.options)
-                    user_input = click.prompt(
-                        clarification.user_guidance + "\nPlease choose a value:\n",
-                        type=choices,
-                    )
-                    workflow = runner.resolve_clarification(clarification, user_input, workflow)
-                if isinstance(clarification, ActionClarification):
-                    webbrowser.open(str(clarification.action_url))
-                    logger().info("Please complete authentication to continue")
-                    workflow = runner.wait_for_ready(workflow)
-                if isinstance(clarification, InputClarification):
-                    user_input = click.prompt(
-                        clarification.user_guidance + "\nPlease enter a value:\n",
-                    )
-                    workflow = runner.resolve_clarification(clarification, user_input, workflow)
-                if isinstance(clarification, ValueConfirmationClarification):
-                    if click.confirm(text=clarification.user_guidance, default=False):
-                        workflow = runner.resolve_clarification(
-                            clarification,
-                            response=True,
-                            workflow=workflow,
-                        )
-                    else:
-                        workflow.state = WorkflowState.FAILED
-                        runner.storage.save_workflow(workflow)
-
-                if isinstance(clarification, CustomClarification):
-                    click.echo(clarification.user_guidance)
-                    click.echo(f"Additional data: {json.dumps(clarification.data)}")
-                    user_input = click.prompt("\nPlease enter a value:\n")
-                    workflow = runner.resolve_clarification(clarification, user_input, workflow)
-
-            runner.execute_workflow(workflow)
-
-        click.echo(workflow.model_dump_json(indent=4))
+        plan_run = portia.run(query)
+        click.echo(plan_run.model_dump_json(indent=4))
 
 
 @click.command()
@@ -238,10 +287,10 @@ def plan(
 ) -> None:
     """Plan a query."""
     cli_config, config = _get_config(**kwargs)
-    runner = Runner(config=config)
+    portia = Portia(config=config)
 
     with execution_context(end_user_id=cli_config.end_user_id):
-        output = runner.generate_plan(query)
+        output = portia.plan(query)
 
     click.echo(output.model_dump_json(indent=4))
 
@@ -297,10 +346,6 @@ def _get_config(
     if len(keys) > 1 and llm_provider is None and llm_model is None:
         message = "Multiple LLM keys found, but no default provided: Select a provider or model"
         raise click.UsageError(message)
-
-    if os.getenv("PORTIA_API_KEY"):
-        config.storage_class = StorageClass.CLOUD
-        config.portia_api_endpoint = os.getenv("PORTIA_API_ENDPOINT") or config.portia_api_endpoint
 
     if llm_provider or llm_model:
         config.llm_provider = llm_provider if llm_provider else llm_model.provider()
