@@ -2,7 +2,7 @@
 
 This agent invokes the OneShotToolCallingModel up to four times, but each individual
 attempt is a one-shot call. It is useful when the tool call is simple, minimizing cost.
-However, for more complex tool calls, the VerifierAgent is recommended as it will
+However, for more complex tool calls, the DefaultExecutionAgent is recommended as it will
 be more successful than the OneShotAgent.
 """
 
@@ -15,19 +15,19 @@ from langchain_core.prompts import ChatPromptTemplate, HumanMessagePromptTemplat
 from langgraph.graph import END, START, MessagesState, StateGraph
 from langgraph.prebuilt import ToolNode
 
-from portia.agents.base_agent import BaseAgent, Output
-from portia.agents.execution_utils import (
+from portia.errors import InvalidAgentError
+from portia.execution_agents.base_execution_agent import BaseExecutionAgent, Output
+from portia.execution_agents.execution_utils import (
     AgentNode,
     next_state_after_tool_call,
     process_output,
     tool_call_or_end,
 )
-from portia.agents.utils.step_summarizer import StepSummarizer
-from portia.errors import InvalidAgentError
+from portia.execution_agents.utils.step_summarizer import StepSummarizer
 from portia.execution_context import get_execution_context
 from portia.llm_wrapper import LLMWrapper
+from portia.plan_run import PlanRun
 from portia.tool import ToolRunContext
-from portia.workflow import Workflow
 
 if TYPE_CHECKING:
     from langchain.tools import StructuredTool
@@ -35,8 +35,8 @@ if TYPE_CHECKING:
 
     from portia.config import Config
     from portia.plan import Step
+    from portia.plan_run import PlanRun
     from portia.tool import Tool
-    from portia.workflow import Workflow
 
 
 class OneShotToolCallingModel:
@@ -47,7 +47,7 @@ class OneShotToolCallingModel:
     are already correctly formatted and complete. This model does not validate
     arguments (e.g., it will not catch missing arguments).
 
-    It is recommended to use the VerifierAgent for more complex tasks.
+    It is recommended to use the DefaultExecutionAgent for more complex tasks.
 
     Args:
         llm (BaseChatModel): The language model to use for generating responses.
@@ -127,7 +127,7 @@ class OneShotToolCallingModel:
         return {"messages": [response]}
 
 
-class OneShotAgent(BaseAgent):
+class OneShotAgent(BaseExecutionAgent):
     """Agent responsible for achieving a task by using langgraph.
 
     This agent performs the following steps:
@@ -136,7 +136,7 @@ class OneShotAgent(BaseAgent):
 
     Args:
         step (Step): The current step in the task plan.
-        workflow (Workflow): The workflow that defines the task execution process.
+        plan_run (PlanRun): The run that defines the task execution process.
         config (Config): The configuration settings for the agent.
         tool (Tool | None): The tool to be used for the task (optional).
 
@@ -148,7 +148,7 @@ class OneShotAgent(BaseAgent):
     def __init__(
         self,
         step: Step,
-        workflow: Workflow,
+        plan_run: PlanRun,
         config: Config,
         tool: Tool | None = None,
     ) -> None:
@@ -156,12 +156,12 @@ class OneShotAgent(BaseAgent):
 
         Args:
             step (Step): The current step in the task plan.
-            workflow (Workflow): The workflow that defines the task execution process.
+            plan_run (PlanRun): The run that defines the task execution process.
             config (Config): The configuration settings for the agent.
             tool (Tool | None): The tool to be used for the task (optional).
 
         """
-        super().__init__(step, workflow, config, tool)
+        super().__init__(step, plan_run, config, tool)
 
     def execute_sync(self) -> Output:
         """Run the core execution logic of the task.
@@ -182,35 +182,35 @@ class OneShotAgent(BaseAgent):
             self.tool.to_langchain_with_artifact(
                 ctx=ToolRunContext(
                     execution_context=get_execution_context(),
-                    workflow_id=self.workflow.id,
+                    plan_run_id=self.plan_run.id,
                     config=self.config,
-                    clarifications=self.workflow.get_clarifications_for_step(),
+                    clarifications=self.plan_run.get_clarifications_for_step(),
                 ),
             ),
         ]
         tool_node = ToolNode(tools)
 
-        workflow = StateGraph(MessagesState)
-        workflow.add_node(
+        graph = StateGraph(MessagesState)
+        graph.add_node(
             AgentNode.TOOL_AGENT,
             OneShotToolCallingModel(llm, context, tools, self).invoke,
         )
-        workflow.add_node(AgentNode.TOOLS, tool_node)
-        workflow.add_node(AgentNode.SUMMARIZER, StepSummarizer(llm).invoke)
-        workflow.add_edge(START, AgentNode.TOOL_AGENT)
+        graph.add_node(AgentNode.TOOLS, tool_node)
+        graph.add_node(AgentNode.SUMMARIZER, StepSummarizer(llm).invoke)
+        graph.add_edge(START, AgentNode.TOOL_AGENT)
 
         # Use execution manager for state transitions
-        workflow.add_conditional_edges(
+        graph.add_conditional_edges(
             AgentNode.TOOL_AGENT,
             tool_call_or_end,
         )
-        workflow.add_conditional_edges(
+        graph.add_conditional_edges(
             AgentNode.TOOLS,
             lambda state: next_state_after_tool_call(state, self.tool),
         )
-        workflow.add_edge(AgentNode.SUMMARIZER, END)
+        graph.add_edge(AgentNode.SUMMARIZER, END)
 
-        app = workflow.compile()
+        app = graph.compile()
         invocation_result = app.invoke({"messages": []})
 
         return process_output(invocation_result["messages"][-1], self.tool)
