@@ -9,11 +9,11 @@ from portia.llm_wrapper import LLMWrapper
 if TYPE_CHECKING:
     from portia.config import Config
     from portia.plan import Plan
-    from portia.workflow import Workflow
+    from portia.plan_run import PlanRun
 
 
 class FinalOutputSummarizer:
-    """Utility class responsible for summarizing the workflow outputs for final output's summary.
+    """Utility class responsible for summarizing the run outputs for final output's summary.
 
     Attributes:
         config (Config): The configuration for the llm.
@@ -38,12 +38,12 @@ class FinalOutputSummarizer:
         """
         self.config = config
 
-    def _build_tasks_and_outputs_context(self, plan: Plan, workflow: Workflow) -> str:
+    def _build_tasks_and_outputs_context(self, plan: Plan, plan_run: PlanRun) -> str:
         """Build the query, tasks and outputs context.
 
         Args:
             plan(Plan): The plan containing the steps.
-            workflow(Workflow): The workflow to get the outputs from.
+            plan_run(PlanRun): The run to get the outputs from.
 
         Returns:
             str: The formatted context string
@@ -53,26 +53,26 @@ class FinalOutputSummarizer:
         context.append(f"Query: {plan.plan_context.query}")
         context.append("----------")
         for step in plan.steps:
-            outputs = workflow.outputs.step_outputs
+            outputs = plan_run.outputs.step_outputs
             if step.output in outputs:
                 context.append(f"Task: {step.task}")
                 context.append(f"Output: {outputs[step.output].value}")
                 context.append("----------")
         return "\n".join(context)
 
-    def create_summary(self, plan: Plan, workflow: Workflow) -> str | None:
+    def create_summary(self, plan: Plan, plan_run: PlanRun) -> str | None:
         """Execute the summarizer llm and return the summary as a string.
 
         Args:
             plan (Plan): The plan containing the steps.
-            workflow (Workflow): The workflow to summarize.
+            plan_run (PlanRun): The run to summarize.
 
         Returns:
             str | None: The generated summary or None if generation fails.
 
         """
         llm = LLMWrapper(self.config).to_langchain()
-        context = self._build_tasks_and_outputs_context(plan, workflow)
+        context = self._build_tasks_and_outputs_context(plan, plan_run)
         response = llm.invoke(
             self.SUMMARIZE_TASK + context,
         )
