@@ -62,11 +62,11 @@ class LLMTool(Tool[str]):
         # Define system and user messages
         context = (
             "Additional context for the LLM tool to use to complete the task, provided by the "
-            "workflow run information and results of other tool calls. Use this to resolve any "
+            "run information and results of other tool calls. Use this to resolve any "
             "tasks"
         )
-        if ctx.execution_context.workflow_run_context:
-            context += f"\nWorkflow run context: {ctx.execution_context.workflow_run_context}"
+        if ctx.execution_context.plan_run_context:
+            context += f"\nRun context: {ctx.execution_context.plan_run_context}"
         if self.tool_context:
             context += f"\nTool context: {self.tool_context}"
         content = task if not len(context.split("\n")) > 1 else f"{context}\n\n{task}"
