@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import json
 import os
 
 import httpx
@@ -11,6 +10,7 @@ from pydantic import BaseModel, Field
 from portia.errors import ToolHardError, ToolSoftError
 from portia.tool import Tool, ToolRunContext
 
+MAX_RESULTS = 3
 
 class SearchToolSchema(BaseModel):
     """Input for SearchTool."""
@@ -30,7 +30,7 @@ class SearchTool(Tool[str]):
     id: str = "search_tool"
     name: str = "Search Tool"
     description: str = (
-        "Searches the internet to find answers to the search query provided and "
+        "Searches the internet (using Tavily) to find answers to the search query provided and "
         "returns those answers, including images, links and a natural language answer. "
         "The search tool has access to general information but can not return specific "
         "information on users or information not available on the internet"
@@ -57,5 +57,6 @@ class SearchTool(Tool[str]):
         response.raise_for_status()
         json_response = response.json()
         if "answer" in json_response:
-            return json.dumps(json_response, indent=2)
+            results = json_response["results"]
+            return results[:MAX_RESULTS]
         raise ToolSoftError(f"Failed to get answer to search: {json_response}")
