@@ -12,7 +12,7 @@ from typing import Any
 from pydantic import BaseModel, ConfigDict
 
 from portia.common import PortiaEnum
-from portia.workflow import WorkflowUUID
+from portia.plan_run import PlanRunUUID
 
 
 class ToolCallStatus(PortiaEnum):
@@ -36,13 +36,13 @@ class ToolCallRecord(BaseModel):
     """Model that records the details of an individual tool call.
 
     This class captures all relevant information about a single tool call
-    within a workflow, including metadata, input and output data, and status.
+    within a PlanRun including metadata, input and output data, and status.
 
     Attributes:
         tool_name (str): The name of the tool being called.
-        workflow_id (WorkflowUUID): The unique identifier of the workflow to which this tool call
+        plan_run_id (RunUUID): The unique identifier of the run to which this tool call
             belongs.
-        step (int): The step number of the tool call in the workflow.
+        step (int): The step number of the tool call in the PlanRun.
         end_user_id (str | None): The ID of the end user, if applicable. Can be None.
         additional_data (dict[str, str]): Additional data from the execution context.
         status (ToolCallStatus): The current status of the tool call (e.g., IN_PROGRESS, SUCCESS).
@@ -55,7 +55,7 @@ class ToolCallRecord(BaseModel):
     model_config = ConfigDict(extra="forbid")
 
     tool_name: str
-    workflow_id: WorkflowUUID
+    plan_run_id: PlanRunUUID
     step: int
     # execution context is tracked here so we get a snapshot if its updated
     end_user_id: str | None
