@@ -7,14 +7,14 @@ from typing import TYPE_CHECKING
 import pytest
 
 from portia.clarification import Clarification, InputClarification
-from portia.config import AgentType, Config, LLMModel, LLMProvider, LogLevel
+from portia.config import Config, ExecutionAgentType, LLMModel, LLMProvider, LogLevel, StorageClass
 from portia.errors import ToolSoftError
 from portia.open_source_tools.registry import example_tool_registry
 from portia.plan import Plan, PlanContext, Step, Variable
-from portia.runner import Runner
+from portia.plan_run import PlanRunState
+from portia.portia import ExecutionHooks, Portia
 from portia.tool_registry import InMemoryToolRegistry
-from portia.workflow import WorkflowState
-from tests.utils import AdditionTool, ClarificationTool, ErrorTool
+from tests.utils import AdditionTool, ClarificationTool, ErrorTool, TestClarificationHandler
 
 if TYPE_CHECKING:
     from portia.tool import ToolRunContext
@@ -35,61 +35,63 @@ PROVIDER_MODELS = [
 ]
 
 AGENTS = [
-    AgentType.VERIFIER,
-    AgentType.ONE_SHOT,
+    ExecutionAgentType.DEFAULT,
+    ExecutionAgentType.ONE_SHOT,
 ]
 
 
 @pytest.mark.parametrize(("llm_provider", "llm_model_name"), PROVIDER_MODELS)
 @pytest.mark.parametrize("agent", AGENTS)
-def test_runner_run_query(
+def test_portia_run_query(
     llm_provider: LLMProvider,
     llm_model_name: LLMModel,
-    agent: AgentType,
+    agent: ExecutionAgentType,
 ) -> None:
-    """Test running a simple query using the Runner."""
+    """Test running a simple query."""
     config = Config.from_default(
         llm_provider=llm_provider,
         llm_model_name=llm_model_name,
-        default_agent_type=agent,
+        execution_agent_type=agent,
+        storage_class=StorageClass.MEMORY,
     )
 
     addition_tool = AdditionTool()
     addition_tool.should_summarize = True
 
     tool_registry = InMemoryToolRegistry.from_local_tools([addition_tool])
-    runner = Runner(config=config, tools=tool_registry)
+    portia = Portia(config=config, tools=tool_registry)
     query = "Add 1 + 2 together"
 
-    workflow = runner.execute_query(query)
+    plan_run = portia.run(query)
 
-    assert workflow.state == WorkflowState.COMPLETE
-    assert workflow.outputs.final_output
-    assert workflow.outputs.final_output.value == 3
-    for output in workflow.outputs.step_outputs.values():
+    assert plan_run.state == PlanRunState.COMPLETE
+    assert plan_run.outputs.final_output
+    assert plan_run.outputs.final_output.value == 3
+    for output in plan_run.outputs.step_outputs.values():
         assert output.summary is not None
 
 
 @pytest.mark.parametrize(("llm_provider", "llm_model_name"), PROVIDER_MODELS)
 @pytest.mark.parametrize("agent", AGENTS)
 @pytest.mark.flaky(reruns=3)
-def test_runner_generate_plan(
+def test_portia_generate_plan(
     llm_provider: LLMProvider,
     llm_model_name: LLMModel,
-    agent: AgentType,
+    agent: ExecutionAgentType,
 ) -> None:
-    """Test planning a simple query using the Runner."""
+    """Test planning a simple query."""
     config = Config.from_default(
         llm_provider=llm_provider,
         llm_model_name=llm_model_name,
-        default_agent_type=agent,
+        execution_agent_type=agent,
+        storage_class=StorageClass.MEMORY,
     )
 
     tool_registry = InMemoryToolRegistry.from_local_tools([AdditionTool()])
-    runner = Runner(config=config, tools=tool_registry)
+    portia = Portia(config=config, tools=tool_registry)
     query = "Add 1 + 2 together"
 
-    plan = runner.generate_plan(query)
+    plan = portia.plan(query)
 
     assert len(plan.steps) == 1
     assert plan.steps[0].tool_id == "add_tool"
@@ -97,31 +99,37 @@ def test_runner_generate_plan(
     assert len(plan.steps[0].inputs) == 2
     assert plan.steps[0].inputs[0].value + plan.steps[0].inputs[1].value == 3
 
-    workflow = runner.execute_query(query)
+    plan_run = portia.run(query)
 
-    assert workflow.state == WorkflowState.COMPLETE
-    assert workflow.outputs.final_output
-    assert workflow.outputs.final_output.value == 3
-    assert workflow.outputs.final_output.summary is not None
+    assert plan_run.state == PlanRunState.COMPLETE
+    assert plan_run.outputs.final_output
+    assert plan_run.outputs.final_output.value == 3
+    assert plan_run.outputs.final_output.summary is not None
 
 
 @pytest.mark.parametrize(("llm_provider", "llm_model_name"), PROVIDER_MODELS)
 @pytest.mark.parametrize("agent", AGENTS)
-def test_runner_run_query_with_clarifications(
+def test_portia_run_query_with_clarifications(
     llm_provider: LLMProvider,
     llm_model_name: LLMModel,
-    agent: AgentType,
+    agent: ExecutionAgentType,
 ) -> None:
-    """Test running a query with clarification using the Runner."""
+    """Test running a query with clarification."""
     config = Config.from_default(
         default_log_level=LogLevel.DEBUG,
         llm_provider=llm_provider,
         llm_model_name=llm_model_name,
-        default_agent_type=agent,
+        execution_agent_type=agent,
+        storage_class=StorageClass.MEMORY,
     )
 
+    test_clarification_handler = TestClarificationHandler()
     tool_registry = InMemoryToolRegistry.from_local_tools([ClarificationTool()])
-    runner = Runner(config=config, tools=tool_registry)
+    portia = Portia(
+        config=config,
+        tools=tool_registry,
+        execution_hooks=ExecutionHooks(clarification_handler=test_clarification_handler),
+    )
     clarification_step = Step(
         tool_id="clarification_tool",
         task="Use tool",
@@ -141,38 +149,79 @@ def test_runner_run_query_with_clarifications(
         ),
         steps=[clarification_step],
     )
-    runner.storage.save_plan(plan)
+    portia.storage.save_plan(plan)
 
-    workflow = runner.create_workflow(plan)
-    workflow = runner.execute_workflow(workflow)
+    plan_run = portia.run_plan(plan)
+    assert plan_run.state == PlanRunState.COMPLETE
+    assert test_clarification_handler.received_clarification is not None
+    assert (
+        test_clarification_handler.received_clarification.user_guidance == "Return a clarification"
+    )
 
-    assert workflow.state == WorkflowState.NEED_CLARIFICATION
-    assert workflow.get_outstanding_clarifications()[0].user_guidance == "Return a clarification"
 
-    workflow = runner.resolve_clarification(
-        workflow.get_outstanding_clarifications()[0],
+def test_portia_run_query_with_clarifications_no_handler() -> None:
+    """Test running a query with clarification using Portia."""
+    config = Config.from_default(
+        default_log_level=LogLevel.DEBUG,
+        llm_provider=LLMProvider.OPENAI,
+        llm_model_name=LLMModel.GPT_4_O_MINI,
+        execution_agent_type=ExecutionAgentType.DEFAULT,
+        storage_class=StorageClass.MEMORY,
+    )
+
+    tool_registry = InMemoryToolRegistry.from_local_tools([ClarificationTool()])
+    portia = Portia(config=config, tools=tool_registry)
+    clarification_step = Step(
+        tool_id="clarification_tool",
+        task="Use tool",
+        output="",
+        inputs=[
+            Variable(
+                name="user_guidance",
+                description="",
+                value="Return a clarification",
+            ),
+        ],
+    )
+    plan = Plan(
+        plan_context=PlanContext(
+            query="raise a clarification",
+            tool_ids=["clarification_tool"],
+        ),
+        steps=[clarification_step],
+    )
+    portia.storage.save_plan(plan)
+
+    plan_run = portia.run_plan(plan)
+
+    assert plan_run.state == PlanRunState.NEED_CLARIFICATION
+    assert plan_run.get_outstanding_clarifications()[0].user_guidance == "Return a clarification"
+
+    plan_run = portia.resolve_clarification(
+        plan_run.get_outstanding_clarifications()[0],
         "False",
     )
 
-    runner.execute_workflow(workflow)
-    assert workflow.state == WorkflowState.COMPLETE
+    portia.resume(plan_run)
+    assert plan_run.state == PlanRunState.COMPLETE
 
 
 @pytest.mark.parametrize(("llm_provider", "llm_model_name"), PROVIDER_MODELS)
 @pytest.mark.parametrize("agent", AGENTS)
-def test_runner_run_query_with_hard_error(
+def test_portia_run_query_with_hard_error(
     llm_provider: LLMProvider,
     llm_model_name: LLMModel,
-    agent: AgentType,
+    agent: ExecutionAgentType,
 ) -> None:
-    """Test running a query with error using the Runner."""
+    """Test running a query with error."""
     config = Config.from_default(
         llm_provider=llm_provider,
         llm_model_name=llm_model_name,
-        default_agent_type=agent,
+        execution_agent_type=agent,
+        storage_class=StorageClass.MEMORY,
     )
     tool_registry = InMemoryToolRegistry.from_local_tools([ErrorTool()])
-    runner = Runner(config=config, tools=tool_registry)
+    portia = Portia(config=config, tools=tool_registry)
     clarification_step = Step(
         tool_id="error_tool",
         task="Use tool",
@@ -202,29 +251,29 @@ def test_runner_run_query_with_hard_error(
         ),
         steps=[clarification_step],
     )
-    runner.storage.save_plan(plan)
-    workflow = runner.create_workflow(plan)
-    workflow = runner.execute_workflow(workflow)
+    portia.storage.save_plan(plan)
+    plan_run = portia.run_plan(plan)
 
-    assert workflow.state == WorkflowState.FAILED
-    assert workflow.outputs.final_output
-    assert isinstance(workflow.outputs.final_output.value, str)
-    assert "Something went wrong" in workflow.outputs.final_output.value
+    assert plan_run.state == PlanRunState.FAILED
+    assert plan_run.outputs.final_output
+    assert isinstance(plan_run.outputs.final_output.value, str)
+    assert "Something went wrong" in plan_run.outputs.final_output.value
 
 
 @pytest.mark.parametrize("agent", AGENTS)
 @pytest.mark.parametrize(("llm_provider", "llm_model_name"), PROVIDER_MODELS)
 @pytest.mark.flaky(reruns=3)
-def test_runner_run_query_with_soft_error(
+def test_portia_run_query_with_soft_error(
     llm_provider: LLMProvider,
     llm_model_name: LLMModel,
-    agent: AgentType,
+    agent: ExecutionAgentType,
 ) -> None:
-    """Test running a query with error using the Runner."""
+    """Test running a query with error."""
     config = Config.from_default(
         llm_provider=llm_provider,
         llm_model_name=llm_model_name,
-        default_agent_type=agent,
+        execution_agent_type=agent,
+        storage_class=StorageClass.MEMORY,
     )
 
     class MyAdditionTool(AdditionTool):
@@ -232,7 +281,7 @@ def test_runner_run_query_with_soft_error(
             raise ToolSoftError("Server Timeout")
 
     tool_registry = InMemoryToolRegistry.from_local_tools([MyAdditionTool()])
-    runner = Runner(config=config, tools=tool_registry)
+    portia = Portia(config=config, tools=tool_registry)
     clarification_step = Step(
         tool_id="add_tool",
         task="Use tool",
@@ -257,44 +306,50 @@ def test_runner_run_query_with_soft_error(
         ),
         steps=[clarification_step],
     )
-    runner.storage.save_plan(plan)
-    workflow = runner.create_workflow(plan)
-    workflow = runner.execute_workflow(workflow)
+    portia.storage.save_plan(plan)
+    plan_run = portia.run_plan(plan)
 
-    assert workflow.state == WorkflowState.FAILED
-    assert workflow.outputs.final_output
-    assert isinstance(workflow.outputs.final_output.value, str)
-    assert "Tool add_tool failed after retries" in workflow.outputs.final_output.value
+    assert plan_run.state == PlanRunState.FAILED
+    assert plan_run.outputs.final_output
+    assert isinstance(plan_run.outputs.final_output.value, str)
+    assert "Tool add_tool failed after retries" in plan_run.outputs.final_output.value
 
 
 @pytest.mark.parametrize(("llm_provider", "llm_model_name"), PROVIDER_MODELS)
 @pytest.mark.parametrize("agent", AGENTS)
 @pytest.mark.flaky(reruns=3)
-def test_runner_run_query_with_multiple_clarifications(
+def test_portia_run_query_with_multiple_clarifications(
     llm_provider: LLMProvider,
     llm_model_name: LLMModel,
-    agent: AgentType,
+    agent: ExecutionAgentType,
 ) -> None:
-    """Test running a query with multiple clarification using the Runner."""
+    """Test running a query with multiple clarification."""
     config = Config.from_default(
         default_log_level=LogLevel.DEBUG,
         llm_provider=llm_provider,
         llm_model_name=llm_model_name,
-        default_agent_type=agent,
+        execution_agent_type=agent,
+        storage_class=StorageClass.MEMORY,
     )
 
     class MyAdditionTool(AdditionTool):
         def run(self, ctx: ToolRunContext, a: int, b: int) -> int | Clarification:  # type: ignore  # noqa: PGH003
             if a == 1:
                 return InputClarification(
-                    workflow_id=ctx.workflow_id,
+                    plan_run_id=ctx.plan_run_id,
                     argument_name="a",
                     user_guidance="please try again",
                 )
             return a + b
 
+    test_clarification_handler = TestClarificationHandler()
+    test_clarification_handler.clarification_response = 456
     tool_registry = InMemoryToolRegistry.from_local_tools([MyAdditionTool()])
-    runner = Runner(config=config, tools=tool_registry)
+    portia = Portia(
+        config=config,
+        tools=tool_registry,
+        execution_hooks=ExecutionHooks(clarification_handler=test_clarification_handler),
+    )
 
     step_one = Step(
         tool_id="add_tool",
@@ -337,34 +392,27 @@ def test_runner_run_query_with_multiple_clarifications(
         ),
         steps=[step_one, step_two],
     )
-    runner.storage.save_plan(plan)
+    portia.storage.save_plan(plan)
 
-    workflow = runner.create_workflow(plan)
-    workflow = runner.execute_workflow(workflow)
+    plan_run = portia.run_plan(plan)
 
-    assert workflow.state == WorkflowState.NEED_CLARIFICATION
-    assert workflow.get_outstanding_clarifications()[0].user_guidance == "please try again"
+    assert plan_run.state == PlanRunState.COMPLETE
+    # 498 = 456 (clarification for value a in step 1) + 2 (value b in step 1)
+    #  + 40 (value b in step 2)
+    assert plan_run.outputs.final_output is not None
+    assert plan_run.outputs.final_output.value == 498
+    assert plan_run.outputs.final_output.summary is not None
 
-    workflow = runner.resolve_clarification(
-        workflow.get_outstanding_clarifications()[0],
-        456,
-        workflow,
-    )
-
-    runner.execute_workflow(workflow)
-    assert workflow.state == WorkflowState.COMPLETE
-    # 498 = 456 (clarification - value a - step 1) + 2 (value b - step 1) + 40 (value b - step 2)
-    assert workflow.outputs.final_output is not None
-    assert workflow.outputs.final_output.value == 498
-    assert workflow.outputs.final_output.summary is not None
+    assert test_clarification_handler.received_clarification is not None
+    assert test_clarification_handler.received_clarification.user_guidance == "please try again"
 
 
-def test_runner_run_query_with_example_registry() -> None:
+def test_portia_run_query_with_example_registry() -> None:
     """Test we can run a query using the example registry."""
     config = Config.from_default()
 
-    runner = Runner(config=config, tools=example_tool_registry)
+    portia = Portia(config=config, tools=example_tool_registry)
     query = "Add 1 + 2 together and then write a haiku about the answer"
 
-    workflow = runner.execute_query(query)
-    assert workflow.state == WorkflowState.COMPLETE
+    plan_run = portia.run(query)
+    assert plan_run.state == PlanRunState.COMPLETE
