@@ -5,12 +5,12 @@ from datetime import UTC, datetime
 import pytest
 from pydantic import HttpUrl
 
-from portia.agents.base_agent import Output
-from portia.agents.context import build_context
 from portia.clarification import ActionClarification, InputClarification
+from portia.execution_agents.base_execution_agent import Output
+from portia.execution_agents.context import build_context
 from portia.execution_context import ExecutionContext
 from portia.plan import Step, Variable
-from tests.utils import get_test_workflow
+from tests.utils import get_test_plan_run
 
 
 @pytest.fixture
@@ -38,11 +38,11 @@ def outputs() -> dict[str, Output]:
 
 def test_context_empty() -> None:
     """Test that the context is set up correctly."""
-    (_, workflow) = get_test_workflow()
+    (_, plan_run) = get_test_plan_run()
     context = build_context(
         ExecutionContext(),
         Step(inputs=[], output="", task=""),
-        workflow,
+        plan_run,
     )
     assert "System Context:" in context
     assert len(context) == 42  # length should always be the same
@@ -50,11 +50,11 @@ def test_context_empty() -> None:
 
 def test_context_execution_context() -> None:
     """Test that the context is set up correctly."""
-    (plan, workflow) = get_test_workflow()
+    (plan, plan_run) = get_test_plan_run()
     context = build_context(
         ExecutionContext(additional_data={"user_id": "123"}),
         plan.steps[0],
-        workflow,
+        plan_run,
     )
     assert "System Context:" in context
     assert "user_id" in context
@@ -63,12 +63,12 @@ def test_context_execution_context() -> None:
 
 def test_context_inputs_only(inputs: list[Variable]) -> None:
     """Test that the context is set up correctly with inputs."""
-    (plan, workflow) = get_test_workflow()
+    (plan, plan_run) = get_test_plan_run()
     plan.steps[0].inputs = inputs
     context = build_context(
         ExecutionContext(),
         plan.steps[0],
-        workflow,
+        plan_run,
     )
     for variable in inputs:
         if variable.value:
@@ -77,13 +77,13 @@ def test_context_inputs_only(inputs: list[Variable]) -> None:
 
 def test_context_inputs_and_outputs(inputs: list[Variable], outputs: dict[str, Output]) -> None:
     """Test that the context is set up correctly with inputs and outputs."""
-    (plan, workflow) = get_test_workflow()
+    (plan, plan_run) = get_test_plan_run()
     plan.steps[0].inputs = inputs
-    workflow.outputs.step_outputs = outputs
+    plan_run.outputs.step_outputs = outputs
     context = build_context(
         ExecutionContext(),
         plan.steps[0],
-        workflow,
+        plan_run,
     )
     for variable in inputs:
         if variable.value:
@@ -96,11 +96,13 @@ def test_context_inputs_and_outputs(inputs: list[Variable], outputs: dict[str, O
 
 def test_system_context() -> None:
     """Test that the system context is set up correctly."""
-    (plan, workflow) = get_test_workflow()
+    (plan, plan_run) = get_test_plan_run()
     context = build_context(
-        ExecutionContext(agent_system_context_extension=["system context 1", "system context 2"]),
+        ExecutionContext(
+            execution_agent_system_context_extension=["system context 1", "system context 2"],
+        ),
         plan.steps[0],
-        workflow,
+        plan_run,
     )
     assert "system context 1" in context
     assert "system context 2" in context
@@ -108,45 +110,45 @@ def test_system_context() -> None:
 
 def test_all_contexts(inputs: list[Variable], outputs: dict[str, Output]) -> None:
     """Test that the context is set up correctly with all contexts."""
-    (plan, workflow) = get_test_workflow()
+    (plan, plan_run) = get_test_plan_run()
     plan.steps[0].inputs = inputs
-    workflow.outputs.step_outputs = outputs
+    plan_run.outputs.step_outputs = outputs
     clarifications = [
         InputClarification(
-            workflow_id=workflow.id,
+            plan_run_id=plan_run.id,
             argument_name="$email_cc",
             user_guidance="email cc list",
             response="bob@bla.com",
             step=0,
         ),
         InputClarification(
-            workflow_id=workflow.id,
+            plan_run_id=plan_run.id,
             argument_name="$email_cc",
             user_guidance="email cc list",
             response="bob@bla.com",
             step=1,
         ),
         ActionClarification(
-            workflow_id=workflow.id,
+            plan_run_id=plan_run.id,
             action_url=HttpUrl("http://example.com"),
             user_guidance="click on the link",
         ),
     ]
-    workflow.outputs.clarifications = clarifications
+    plan_run.outputs.clarifications = clarifications
     context = build_context(
         ExecutionContext(
-            agent_system_context_extension=["system context 1", "system context 2"],
+            execution_agent_system_context_extension=["system context 1", "system context 2"],
             end_user_id="123",
             additional_data={"email": "hello@world.com"},
         ),
         plan.steps[0],
-        workflow,
+        plan_run,
     )
     # as LLMs are sensitive even to white space formatting we do a complete match here
     assert (
         context
         == f"""Additional context: You MUST use this information to complete your task.
-Inputs: the original inputs provided by the planner
+Inputs: the original inputs provided by the planning_agent
 input_name: $email_address
 input_value: test@example.com
 input_description: Target recipient for email
@@ -186,29 +188,31 @@ def test_context_inputs_outputs_clarifications(
     outputs: dict[str, Output],
 ) -> None:
     """Test that the context is set up correctly with inputs, outputs, and missing args."""
-    (plan, workflow) = get_test_workflow()
+    (plan, plan_run) = get_test_plan_run()
     clarifications = [
         InputClarification(
-            workflow_id=workflow.id,
+            plan_run_id=plan_run.id,
             argument_name="$email_cc",
             user_guidance="email cc list",
             response="bob@bla.com",
             step=0,
         ),
         ActionClarification(
-            workflow_id=workflow.id,
+            plan_run_id=plan_run.id,
             action_url=HttpUrl("http://example.com"),
             user_guidance="click on the link",
             step=1,
         ),
     ]
     plan.steps[0].inputs = inputs
-    workflow.outputs.step_outputs = outputs
-    workflow.outputs.clarifications = clarifications
+    plan_run.outputs.step_outputs = outputs
+    plan_run.outputs.clarifications = clarifications
     context = build_context(
-        ExecutionContext(agent_system_context_extension=["system context 1", "system context 2"]),
+        ExecutionContext(
+            execution_agent_system_context_extension=["system context 1", "system context 2"],
+        ),
         plan.steps[0],
-        workflow,
+        plan_run,
     )
     for variable in inputs:
         if variable.value:
