@@ -2,16 +2,16 @@
 
 from unittest import mock
 
-from portia.agents.base_agent import Output
-from portia.agents.utils.final_output_summarizer import FinalOutputSummarizer
+from portia.execution_agents.base_execution_agent import Output
+from portia.execution_agents.utils.final_output_summarizer import FinalOutputSummarizer
 from portia.plan import Step
-from tests.utils import get_test_config, get_test_workflow
+from tests.utils import get_test_config, get_test_plan_run
 
 
 def test_summarizer_agent_execute_sync() -> None:
     """Test that the summarizer agent correctly executes and returns a summary."""
     # Set up test data
-    (plan, workflow) = get_test_workflow()
+    (plan, plan_run) = get_test_plan_run()
     plan.plan_context.query = "What's the weather in London and what can I do?"
     plan.steps = [
         Step(
@@ -24,7 +24,7 @@ def test_summarizer_agent_execute_sync() -> None:
         ),
     ]
 
-    workflow.outputs.step_outputs = {
+    plan_run.outputs.step_outputs = {
         "$london_weather": Output(value="Sunny and warm"),
         "$activities": Output(value="Visit Hyde Park and have a picnic"),
     }
@@ -34,11 +34,13 @@ def test_summarizer_agent_execute_sync() -> None:
     mock_llm = mock.MagicMock()
     mock_llm.invoke.return_value.content = expected_summary
 
-    with mock.patch("portia.agents.utils.final_output_summarizer.LLMWrapper") as mock_wrapper:
+    with mock.patch(
+        "portia.execution_agents.utils.final_output_summarizer.LLMWrapper",
+    ) as mock_wrapper:
         mock_wrapper.return_value.to_langchain.return_value = mock_llm
 
         summarizer = FinalOutputSummarizer(config=get_test_config())
-        output = summarizer.create_summary(plan=plan, workflow=workflow)
+        output = summarizer.create_summary(plan=plan, plan_run=plan_run)
 
         assert output == expected_summary
 
@@ -57,21 +59,23 @@ def test_summarizer_agent_execute_sync() -> None:
         mock_llm.invoke.assert_called_once_with(expected_prompt)
 
 
-def test_summarizer_agent_empty_workflow() -> None:
-    """Test summarizer agent with empty workflow."""
-    (plan, workflow) = get_test_workflow()
+def test_summarizer_agent_empty_plan_run() -> None:
+    """Test summarizer agent with empty plan run."""
+    (plan, plan_run) = get_test_plan_run()
     plan.plan_context.query = "Empty query"
     plan.steps = []
-    workflow.outputs.step_outputs = {}
+    plan_run.outputs.step_outputs = {}
 
     mock_llm = mock.MagicMock()
     mock_llm.invoke.return_value.content = "Empty summary"
 
-    with mock.patch("portia.agents.utils.final_output_summarizer.LLMWrapper") as mock_wrapper:
+    with mock.patch(
+        "portia.execution_agents.utils.final_output_summarizer.LLMWrapper",
+    ) as mock_wrapper:
         mock_wrapper.return_value.to_langchain.return_value = mock_llm
 
         summarizer = FinalOutputSummarizer(config=get_test_config())
-        output = summarizer.create_summary(plan=plan, workflow=workflow)
+        output = summarizer.create_summary(plan=plan, plan_run=plan_run)
 
         # Verify empty context case
         assert output == "Empty summary"
@@ -81,17 +85,19 @@ def test_summarizer_agent_empty_workflow() -> None:
 
 def test_summarizer_agent_handles_none_response() -> None:
     """Test that the agent handles None response from LLM."""
-    (plan, workflow) = get_test_workflow()
+    (plan, plan_run) = get_test_plan_run()
     plan.plan_context.query = "Test query"
 
     mock_llm = mock.MagicMock()
     mock_llm.invoke.return_value.content = None
 
-    with mock.patch("portia.agents.utils.final_output_summarizer.LLMWrapper") as mock_wrapper:
+    with mock.patch(
+        "portia.execution_agents.utils.final_output_summarizer.LLMWrapper",
+    ) as mock_wrapper:
         mock_wrapper.return_value.to_langchain.return_value = mock_llm
 
         summarizer = FinalOutputSummarizer(config=get_test_config())
-        output = summarizer.create_summary(plan=plan, workflow=workflow)
+        output = summarizer.create_summary(plan=plan, plan_run=plan_run)
 
         # Verify None handling
         assert output is None
@@ -99,7 +105,7 @@ def test_summarizer_agent_handles_none_response() -> None:
 
 def test_build_tasks_and_outputs_context() -> None:
     """Test that the tasks and outputs context is built correctly."""
-    (plan, workflow) = get_test_workflow()
+    (plan, plan_run) = get_test_plan_run()
 
     # Set up test data
     plan.plan_context.query = "What's the weather in London and what can I do?"
@@ -114,7 +120,7 @@ def test_build_tasks_and_outputs_context() -> None:
         ),
     ]
 
-    workflow.outputs.step_outputs = {
+    plan_run.outputs.step_outputs = {
         "$london_weather": Output(value="Sunny and warm"),
         "$activities": Output(value="Visit Hyde Park and have a picnic"),
     }
@@ -122,7 +128,7 @@ def test_build_tasks_and_outputs_context() -> None:
     summarizer = FinalOutputSummarizer(config=get_test_config())
     context = summarizer._build_tasks_and_outputs_context(  # noqa: SLF001
         plan=plan,
-        workflow=workflow,
+        plan_run=plan_run,
     )
 
     # Verify exact output format including query
@@ -140,17 +146,17 @@ def test_build_tasks_and_outputs_context() -> None:
 
 def test_build_tasks_and_outputs_context_empty() -> None:
     """Test that the tasks and outputs context handles empty steps and outputs."""
-    (plan, workflow) = get_test_workflow()
+    (plan, plan_run) = get_test_plan_run()
 
-    # Empty plan and workflow
+    # Empty plan and run
     plan.plan_context.query = "Empty query"
     plan.steps = []
-    workflow.outputs.step_outputs = {}
+    plan_run.outputs.step_outputs = {}
 
     summarizer = FinalOutputSummarizer(config=get_test_config())
     context = summarizer._build_tasks_and_outputs_context(  # noqa: SLF001
         plan=plan,
-        workflow=workflow,
+        plan_run=plan_run,
     )
 
     # Should still include query even if no steps/outputs
@@ -159,7 +165,7 @@ def test_build_tasks_and_outputs_context_empty() -> None:
 
 def test_build_tasks_and_outputs_context_partial_outputs() -> None:
     """Test that the context builder handles steps with missing outputs."""
-    (plan, workflow) = get_test_workflow()
+    (plan, plan_run) = get_test_plan_run()
 
     # Set up test data with missing output
     plan.plan_context.query = "What's the weather in London?"
@@ -175,14 +181,14 @@ def test_build_tasks_and_outputs_context_partial_outputs() -> None:
     ]
 
     # Only provide output for first step
-    workflow.outputs.step_outputs = {
+    plan_run.outputs.step_outputs = {
         "$london_weather": Output(value="Sunny and warm"),
     }
 
     summarizer = FinalOutputSummarizer(config=get_test_config())
     context = summarizer._build_tasks_and_outputs_context(  # noqa: SLF001
         plan=plan,
-        workflow=workflow,
+        plan_run=plan_run,
     )
 
     # Verify only step with output is included, but query is always present
