@@ -26,7 +26,7 @@ def mock_llm_tool() -> LLMTool:
 
 @patch("portia.open_source_tools.llm_tool.LLMWrapper")
 @patch.dict(os.environ, {"OPENAI_API_KEY": "123"})
-def test_llm_tool_run(
+def test_llm_tool_plan_run(
     mock_llm_wrapper: MagicMock,
     mock_execution_context: MagicMock,
     mock_llm_tool: MagicMock,
@@ -40,8 +40,8 @@ def test_llm_tool_run(
     mock_llm_wrapper.return_value.to_langchain.return_value = mock_llm
     mock_execution_context.execution_context = MagicMock()
     mock_execution_context.config = get_test_config()
-    mock_execution_context.workflow_id = uuid.uuid4()
-    mock_execution_context.execution_context.workflow_run_context = None
+    mock_execution_context.plan_run_id = uuid.uuid4()
+    mock_execution_context.execution_context.plan_run_context = None
     # Define task input
     task = "What is the capital of France?"
 
@@ -92,7 +92,7 @@ def test_llm_tool_run_with_context(
     mock_llm_wrapper.return_value.to_langchain.return_value = mock_llm
     mock_execution_context.execution_context = MagicMock()
     mock_execution_context.config = get_test_config()
-    mock_execution_context.workflow_id = uuid.uuid4()
+    mock_execution_context.plan_run_id = uuid.uuid4()
     # Define task and context
     mock_llm_tool.tool_context = "Context for task"
     task = "What is the capital of France?"
