@@ -1,6 +1,5 @@
 """Search tool tests."""
 
-import json
 from unittest.mock import Mock, patch
 
 import pytest
@@ -28,7 +27,7 @@ def test_search_tool_successful_response() -> None:
         "follow_up_questions": "",
         "answer": "The capital of France is Paris.",
         "images": [],
-        "results": [],
+        "results": ["result1", "result2", "result3", "result4", "result5"],
         "response_time": 2.43,
     }
 
@@ -38,8 +37,28 @@ def test_search_tool_successful_response() -> None:
             mock_post.return_value = Mock(status_code=200, json=lambda: mock_response)
 
             result = tool.run(ctx, "What is the capital of France?")
-            assert result == json.dumps(mock_response, indent=2)
+            assert result == ["result1", "result2", "result3"]
 
+def test_search_tool_fewer_results_than_max() -> None:
+    """Test that SearchTool successfully processes a valid response."""
+    tool = SearchTool()
+    mock_api_key = "mock-api-key"
+    mock_response = {
+        "query": "What is the capital of France?",
+        "follow_up_questions": "",
+        "answer": "The capital of France is Paris.",
+        "images": [],
+        "results": ["result1", "result2"],
+        "response_time": 2.43,
+    }
+
+    with patch("os.getenv", return_value=mock_api_key):
+        ctx = get_test_tool_context()
+        with patch("httpx.post") as mock_post:
+            mock_post.return_value = Mock(status_code=200, json=lambda: mock_response)
+
+            result = tool.run(ctx, "What is the capital of France?")
+            assert result == ["result1", "result2"]
 
 def test_search_tool_no_answer_in_response() -> None:
     """Test that SearchTool raises ToolSoftError if no answer is found in the response."""
