@@ -13,9 +13,9 @@ from portia.clarification import (
     CustomClarification,
     MultipleChoiceClarification,
 )
-from portia.prefixed_uuid import WorkflowUUID
+from portia.prefixed_uuid import PlanRunUUID
 from portia.storage import DiskFileStorage
-from tests.utils import get_test_workflow
+from tests.utils import get_test_plan_run
 
 if TYPE_CHECKING:
     from pathlib import Path
@@ -24,7 +24,7 @@ if TYPE_CHECKING:
 def test_action_clarification_ser() -> None:
     """Test action clarifications can be serialized."""
     clarification = ActionClarification(
-        workflow_id=WorkflowUUID(),
+        plan_run_id=PlanRunUUID(),
         user_guidance="test",
         action_url=HttpUrl("https://example.com"),
     )
@@ -35,7 +35,7 @@ def test_action_clarification_ser() -> None:
 def test_clarification_uuid_assign() -> None:
     """Test clarification assign correct UUIDs."""
     clarification = ActionClarification(
-        workflow_id=WorkflowUUID(),
+        plan_run_id=PlanRunUUID(),
         user_guidance="test",
         action_url=HttpUrl("https://example.com"),
     )
@@ -46,7 +46,7 @@ def test_value_multi_choice_validation() -> None:
     """Test clarifications error on invalid response."""
     with pytest.raises(ValueError):  # noqa: PT011
         MultipleChoiceClarification(
-            workflow_id=WorkflowUUID(),
+            plan_run_id=PlanRunUUID(),
             argument_name="test",
             user_guidance="test",
             options=["yes"],
@@ -55,7 +55,7 @@ def test_value_multi_choice_validation() -> None:
         )
 
     MultipleChoiceClarification(
-        workflow_id=WorkflowUUID(),
+        plan_run_id=PlanRunUUID(),
         argument_name="test",
         user_guidance="test",
         options=["yes"],
@@ -66,10 +66,10 @@ def test_value_multi_choice_validation() -> None:
 
 def test_custom_clarification_deserialize(tmp_path: Path) -> None:
     """Test clarifications error on invalid response."""
-    (plan, workflow) = get_test_workflow()
+    (plan, plan_run) = get_test_plan_run()
 
     clarification_one = CustomClarification(
-        workflow_id=workflow.id,
+        plan_run_id=plan_run.id,
         user_guidance="Please provide data",
         name="My Clarification",
         data={"email": {"test": "hello@example.com"}},
@@ -77,10 +77,10 @@ def test_custom_clarification_deserialize(tmp_path: Path) -> None:
 
     storage = DiskFileStorage(storage_dir=str(tmp_path))
 
-    workflow.outputs.clarifications = [clarification_one]
+    plan_run.outputs.clarifications = [clarification_one]
 
     storage.save_plan(plan)
-    storage.save_workflow(workflow)
-    retrieved = storage.get_workflow(workflow.id)
+    storage.save_plan_run(plan_run)
+    retrieved = storage.get_plan_run(plan_run.id)
     assert isinstance(retrieved.outputs.clarifications[0], CustomClarification)
     assert retrieved.outputs.clarifications[0].data == {"email": {"test": "hello@example.com"}}
