@@ -128,7 +128,7 @@ def test_remote_tool_hard_error_from_server() -> None:
             "arguments": {},
             "execution_context": {
                 "end_user_id": ctx.execution_context.end_user_id or "",
-                "workflow_id": str(ctx.workflow_id),
+                "plan_run_id": str(ctx.plan_run_id),
                 "additional_data": ctx.execution_context.additional_data or {},
             },
         }
@@ -172,7 +172,7 @@ def test_remote_tool_soft_error() -> None:
             "arguments": {},
             "execution_context": {
                 "end_user_id": ctx.execution_context.end_user_id or "",
-                "workflow_id": str(ctx.workflow_id),
+                "plan_run_id": str(ctx.plan_run_id),
                 "additional_data": ctx.execution_context.additional_data or {},
             },
         }
@@ -215,7 +215,7 @@ def test_remote_tool_bad_response() -> None:
             "arguments": {},
             "execution_context": {
                 "end_user_id": ctx.execution_context.end_user_id or "",
-                "workflow_id": str(ctx.workflow_id),
+                "plan_run_id": str(ctx.plan_run_id),
                 "additional_data": ctx.execution_context.additional_data or {},
             },
         }
@@ -259,7 +259,7 @@ def test_remote_tool_hard_error() -> None:
             "arguments": {},
             "execution_context": {
                 "end_user_id": ctx.execution_context.end_user_id or "",
-                "workflow_id": str(ctx.workflow_id),
+                "plan_run_id": str(ctx.plan_run_id),
                 "additional_data": ctx.execution_context.additional_data or {},
             },
         }
@@ -299,7 +299,7 @@ def test_remote_tool_ready() -> None:
         content = {
             "execution_context": {
                 "end_user_id": ctx.execution_context.end_user_id or "",
-                "workflow_id": str(ctx.workflow_id),
+                "plan_run_id": str(ctx.plan_run_id),
                 "additional_data": ctx.execution_context.additional_data or {},
             },
         }
@@ -343,7 +343,7 @@ def test_remote_tool_ready_error() -> None:
         content = {
             "execution_context": {
                 "end_user_id": ctx.execution_context.end_user_id or "",
-                "workflow_id": str(ctx.workflow_id),
+                "plan_run_id": str(ctx.plan_run_id),
                 "additional_data": ctx.execution_context.additional_data or {},
             },
         }
@@ -399,7 +399,7 @@ def test_remote_tool_action_clarifications() -> None:
             "arguments": {},
             "execution_context": {
                 "end_user_id": ctx.execution_context.end_user_id or "",
-                "workflow_id": str(ctx.workflow_id),
+                "plan_run_id": str(ctx.plan_run_id),
                 "additional_data": ctx.execution_context.additional_data or {},
             },
         }
@@ -455,7 +455,7 @@ def test_remote_tool_input_clarifications() -> None:
             "arguments": {},
             "execution_context": {
                 "end_user_id": ctx.execution_context.end_user_id or "",
-                "workflow_id": str(ctx.workflow_id),
+                "plan_run_id": str(ctx.plan_run_id),
                 "additional_data": ctx.execution_context.additional_data or {},
             },
         }
@@ -513,7 +513,7 @@ def test_remote_tool_mc_clarifications() -> None:
             "arguments": {},
             "execution_context": {
                 "end_user_id": ctx.execution_context.end_user_id or "",
-                "workflow_id": str(ctx.workflow_id),
+                "plan_run_id": str(ctx.plan_run_id),
                 "additional_data": ctx.execution_context.additional_data or {},
             },
         }
@@ -569,7 +569,7 @@ def test_remote_tool_value_confirm_clarifications() -> None:
             "arguments": {},
             "execution_context": {
                 "end_user_id": ctx.execution_context.end_user_id or "",
-                "workflow_id": str(ctx.workflow_id),
+                "plan_run_id": str(ctx.plan_run_id),
                 "additional_data": ctx.execution_context.additional_data or {},
             },
         }
